refactor(storage): extract registry_mirror module (R.6)
Phase R.6 of the storage modularization plan
(SPEC_STORE_MODULARIZATION_2026_05_27.md). Moves the cross-version
JSON-registry mirror out of store.rs (4975 → 4843 lines) into a new
`registry_mirror.rs` (176 LOC). Bonus: this whole module retires
when Phase R sunsets the JSON registry — isolating it now makes
that future deletion a clean unit.

What moves:
- `Store::registry_upsert_if_named` (pub(super) — only mutation paths
  in store.rs call it)
- File-level helpers `agent_instance_to_record`, `empty_to_none`,
  `relative_workdir`

What stays:
- `Store::set_registry`, `Store::registry()` private accessor, and
  `Store::shared_agent_registry()` public accessor — these are Store
  lifecycle, not mirror logic.

The mirror method changed from `self.registry()` (private accessor)
to `self.shared_agent_registry()` (public) since the registry()
helper isn't visible across modules. Functionally identical.

Net store.rs change: -132 lines. New file 176 LOC.

162/162 storage tests green.
