+++ .changesets/1779975470-refactor-storage-extract-registry-mirror-module-r-6-l18j.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(storage): extract registry_mirror module (R.6)
