+++ agentmux-srv/src/backend/storage/mod.rs
@@ -13,6 +13,7 @@ pub mod history;
 pub mod identities;
 pub mod memory_bundles;
 pub mod migrations;
+pub mod registry_mirror;
 pub mod skills;
 pub mod snapshot;
 pub mod store;
